Move to a promises-based architecture.
I'm not a super huge fan of native promises (I'd rather call a .resolve
method on a promise object), but I prefer native to requiring external
dependencies.

What's key here is that getScene now returns a Promise, which resolves
with the appropriate data when the inner callback on
Three.TextureLoader.load finishes. The calling code needs to now use
.then(function (someData) { ... }) instead of direct assignment-and-use.

This also includes some incidental linting.